+++ activerecord/test/cases/migration/references_foreign_key_test.rb
@@ -144,6 +144,22 @@ module ActiveRecord
           @connection.drop_table "testing", if_exists: true
         end
       end
+
+      test "multiple foreign keys can be added to the same table" do
+        @connection.create_table :testings do |t|
+          t.integer :col_1
+          t.integer :col_2
+
+          t.foreign_key :testing_parents, column: :col_1
+          t.foreign_key :testing_parents, column: :col_2
+        end
+
+        fks = @connection.foreign_keys("testings")
+
+        fk_definitions = fks.map {|fk| [fk.from_table, fk.to_table, fk.column] }
+        assert_equal([["testings", "testing_parents", "col_1"],
+                      ["testings", "testing_parents", "col_2"]], fk_definitions)
+      end
     end
   end
 end